Add explicit get list items count query to ListItemSqlUtils

+++ libs/fluxc/src/main/java/org/wordpress/android/fluxc/model/list/datastore/ListDataStoreInterface.kt
@@ -14,7 +14,7 @@ interface ListDataStoreInterface<T> {
     /**
      * Should fetch the list for the given [ListDescriptor] and an offset.
      */
-    fun fetchList(listDescriptor: ListDescriptor, offset: Int)
+    fun fetchList(listDescriptor: ListDescriptor, offset: Long)
 
     /**
      * Should return the item for the given [ListDescriptor] and [remoteItemId] or null if it's not available.
